New changes to Blake
I have finished the initial translation to java, I have no idea how it
is going to work as of yet. Would appreciate second eyes on it.

+++ blake/BLAKEHashState.java
@@ -10,9 +10,11 @@
 
 package blake;
 
+import java.math.BigInteger;
+
 public class BLAKEHashState {
     public short hashbitlen;
-    public int datalen;
+    public long datalen;
     public short init;
     public short nullt;
     public int h32[];
